+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.0-rc.1] - 2021-12-11
+
+* Null safety
+* Support for notus 1.0.0-rc.1
+
+## [0.0.9-dev.3.0] - 2021-03-05
+
+* Fixed a bug where embeds were positioned incorrectly
+* Support the new underline and strikethrough features from Zefyr
+
+## [0.0.9-dev.2.0] - 2021-02-14
+
+* Recognise embeds when they appear as Map
+
+## [0.0.9-dev.1.0] - 2021-02-14
+
+* Support notus 1.0.0-dev versions and quill delta 2.0.0
+
 ## [0.0.8] - 2020-11-04
 
 * Support decoding of `<p>` elements with children